Boundary triangulations working except for normal vector
