+++ test/new/GeometryTests/runtests.jl
@@ -24,4 +24,8 @@ using Test
 
 @testset "CartesianDiscreteModels" begin include("CartesianDiscreteModelsTests.jl") end
 
+@testset "BoundaryTriangulations" begin include("BoundaryTriangulationsTests.jl") end
+
+@testset "GenericBoundaryTriangulations" begin include("GenericBoundaryTriangulationsTests.jl") end
+
 end # module
